Move serialization to a middleware
