+++ tinydb/database.py
@@ -30,6 +30,8 @@ class TinyDB(object):
     and getting tables.
     """
 
+    DEFAULT_STORAGE = JSONStorage
+
     def __init__(self, *args, **kwargs):
         """
         Create a new instance of TinyDB.
@@ -40,10 +42,9 @@ class TinyDB(object):
         :param storage: The class of the storage to use. Will be initialized
                         with ``args`` and ``kwargs``.
         """
-        storage = kwargs.pop('storage', JSONStorage)
+        storage = kwargs.pop('storage', TinyDB.DEFAULT_STORAGE)
         #: :type: Storage
         self._storage = storage(*args, **kwargs)
-        self._serializers = {}
 
         self._table_cache = {}
         self._table = self.table('_default')
@@ -102,24 +103,6 @@ class TinyDB(object):
         self._write({})
         self._table_cache.clear()
 
-    def register_serializer(self, serializer, name):
-        """
-        Register a new Serializer.
-
-        When reading from/writing to the underlying storage, TinyDB
-        will run all objects through the list of registered serializers
-        allowing each one to handle objects it recognizes.
-
-        .. note:: The name has to be unique among this database instance.
-                  Re-using the same name will overwrite the old serializer.
-                  Also, registering a serializer will be reflected in all
-                  tables when reading/writing them.
-
-        :param serializer: an instance of the serializer
-        :type serializer: tinydb.serialize.Serializer
-        """
-        self._serializers[name] = serializer
-
     def _read(self):
         """
         Reading access to the backend.
@@ -142,31 +125,9 @@ class TinyDB(object):
         """
 
         try:
-            return self._deserialize(self._read()[table])
+            return self._read()[table]
         except (KeyError, TypeError):
             return {}
-
-    def _deserialize(self, data):
-        """
-        Deserialize the data passed in with all registered serializers
-
-        :param data: the data set to deserialize
-        :return: the data set with deserialized values as needed
-        """
-        for serializer_name in self._serializers:
-            serializer = self._serializers[serializer_name]
-            tag = '{{{}}}:'.format(serializer_name)  # E.g: '{TinyDate}:'
-
-            for eid in data:
-                for field in data[eid]:
-                    try:
-                        if data[eid][field].startswith(tag):
-                            encoded = data[eid][field][len(tag):]
-                            data[eid][field] = serializer.decode(encoded)
-                    except AttributeError:
-                        pass  # Not a string
-
-        return data
 
     def _write(self, tables):
         """
@@ -193,36 +154,7 @@ class TinyDB(object):
         data = self._read()
         data[table] = values
 
-        # Finally, serialize & store the data
-        self._write(self._serialize(data))
-
-    def _serialize(self, data):
-        """
-        Serialize the data passed in with all registered serializers
-
-        :param data: the data set to serialize
-        :return: the data set with serialized values as needed
-        """
-
-        for serializer_name in self._serializers:
-            # If no serializers are registered, this code will just look up
-            # the serializer list and continue. But if there are serializers,
-            # the inner loop will run very often.
-            # For that reason, the lookup of the serialized class is pulled
-            # out into the outer loop:
-
-            serializer = self._serializers[serializer_name]
-            serializer_class = serializer.OBJ_CLASS
-
-            for eid in data:
-                for field in data[eid]:
-                    if isinstance(data[eid][field], serializer_class):
-                        encoded = serializer.encode(data[eid][field])
-                        tagged = '{{{}}}:{}'.format(serializer_name, encoded)
-
-                        data[eid][field] = tagged
-
-        return data
+        self._write(data)
 
     # Methods that are executed on the default table
     # Because magic methods are not handlet by __getattr__ we need to forward
